# openapiconfig-privsense.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "PrivSense API",
      "description": "API for detecting personally identifiable information (PII) in relational databases",
      "contact": {
        "name": "PrivSense Team",
        "url": "https://www.privsense.com",
        "email": "support@privsense.com"
      },
      "license": {
        "name": "Proprietary",
        "url": "https://www.privsense.com/license"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/privsense",
        "description": "Default server URL"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
...
    "paths": {
...
      "/api/v1/config": {
        "get": {
          "tags": [
            "Configuration"
          ],
          "summary": "Get application configuration",
          "description": "Returns the current application configuration",
          "operationId": "getConfiguration",
          "responses": {
            "200": {
              "description": "Configuration retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ConfigurationDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Configuration"
          ],
          "summary": "Update application configuration",
          "description": "Updates the application configuration",
          "operationId": "updateConfiguration",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Configuration updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ConfigurationDTO"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid configuration",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ConfigurationDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/config/detection-rules/{ruleId}": {
        "get": {
          "tags": [
            "Configuration"
          ],
          "summary": "Get detection rule",
          "description": "Returns a specific detection rule",
          "operationId": "getDetectionRule",
          "parameters": [
            {
              "name": "ruleId",
              "in": "path",
              "description": "ID of the rule",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Rule retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DetectionRuleDTO"
                  }
                }
              }
            },
            "404": {
              "description": "Rule not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DetectionRuleDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Configuration"
          ],
          "summary": "Update detection rule",
          "description": "Updates an existing detection rule",
          "operationId": "updateDetectionRule",
          "parameters": [
            {
              "name": "ruleId",
              "in": "path",
              "description": "ID of the rule",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionRuleDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "400": {
              "description": "Invalid rule data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DetectionRuleDTO"
                  }
                }
              }
            },
            "200": {
              "description": "Rule updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DetectionRuleDTO"
                  }
                }
              }
            },
            "404": {
              "description": "Rule not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DetectionRuleDTO"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Configuration"
          ],
          "summary": "Delete detection rule",
          "description": "Deletes a detection rule",
          "operationId": "deleteDetectionRule",
          "parameters": [
            {
              "name": "ruleId",
              "in": "path",
              "description": "ID of the rule",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Rule deleted successfully"
            },
            "404": {
              "description": "Rule not found"
            }
          }
        }
      },
...
      "/api/v1/sampling": {
        "post": {
          "tags": [
            "Database Sampling"
          ],
          "summary": "Test column sampling",
          "description": "Samples data from a specific column to validate sampling functionality",
          "operationId": "testSampling",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplingRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "400": {
              "description": "Invalid request parameters",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SamplingResponse"
                  }
                }
              }
            },
            "200": {
              "description": "Sampling successful",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SamplingResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Database connection or column not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SamplingResponse"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/config/detection-rules": {
        "get": {
          "tags": [
            "Configuration"
          ],
          "summary": "Get all detection rules",
          "description": "Returns all detection rules",
          "operationId": "getDetectionRules",
          "responses": {
            "200": {
              "description": "Rules retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DetectionRuleDTO"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Configuration"
          ],
          "summary": "Create detection rule",
          "description": "Creates a new detection rule",
          "operationId": "createDetectionRule",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionRuleDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "400": {
              "description": "Invalid rule data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DetectionRuleDTO"
                  }
                }
              }
            },
            "201": {
              "description": "Rule created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DetectionRuleDTO"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "DetectionConfigDTO": {
          "type": "object",
          "properties": {
            "heuristicThreshold": {
              "type": "number",
              "description": "Threshold for heuristic-based detection",
              "format": "double",
              "example": 0.7
            },
            "regexThreshold": {
              "type": "number",
              "description": "Threshold for regex-based detection",
              "format": "double",
              "example": 0.8
            },
            "nerThreshold": {
              "type": "number",
              "description": "Threshold for NER-based detection",
              "format": "double",
              "example": 0.6
            },
            "reportingThreshold": {
              "type": "number",
              "description": "Minimum threshold for reporting PII findings",
              "format": "double",
              "example": 0.5
            },
            "stopPipelineOnHighConfidence": {
              "type": "boolean",
              "description": "Whether to stop detection pipeline on high confidence matches"
            },
            "entropyEnabled": {
              "type": "boolean",
              "description": "Whether entropy-based detection is enabled"
            }
          },
          "description": "PII detection configuration settings"
        },
        "SamplingConfigDTO": {
          "type": "object",
          "properties": {
            "defaultSize": {
              "type": "integer",
              "description": "Default sample size",
              "format": "int32",
              "example": 100
            },
            "maxConcurrentDbQueries": {
              "type": "integer",
              "description": "Maximum concurrent database queries",
              "format": "int32",
              "example": 10
            },
            "entropyCalculationEnabled": {
              "type": "boolean",
              "description": "Whether entropy calculation is enabled during sampling"
            },
            "defaultMethod": {
              "type": "string",
              "description": "Default sampling method",
              "example": "RANDOM"
            }
          },
          "description": "Data sampling configuration settings"
        },
...
        "CircuitBreakerConfigDTO": {
          "type": "object",
          "properties": {
            "enabled": {
              "type": "boolean",
              "description": "Whether circuit breaker is enabled"
            },
            "failureThreshold": {
              "type": "integer",
              "description": "Failure threshold before opening circuit",
              "format": "int32",
              "example": 5
            },
            "resetTimeoutSeconds": {
              "type": "integer",
              "description": "Reset timeout in seconds",
              "format": "int32",
              "example": 30
            }
          },
          "description": "Circuit breaker configuration settings"
        },
        "ConfigurationDTO": {
          "type": "object",
          "properties": {
            "detection": {
              "$ref": "#/components/schemas/DetectionConfigDTO"
            },
            "sampling": {
              "$ref": "#/components/schemas/SamplingConfigDTO"
            },
            "ner": {
              "$ref": "#/components/schemas/NerConfigDTO"
            },
            "reporting": {
              "$ref": "#/components/schemas/ReportingConfigDTO"
            },
            "database": {
              "$ref": "#/components/schemas/DatabaseConfigDTO"
            }
          },
          "description": "Application configuration settings"
        },
        "DatabaseConfigDTO": {
          "type": "object",
          "properties": {
            "pool": {
              "$ref": "#/components/schemas/PoolConfigDTO"
            },
            "driverDir": {
              "type": "string",
              "description": "Directory for JDBC drivers",
              "example": "./drivers"
            }
          },
          "description": "Database configuration settings"
        },
        "NerConfigDTO": {
          "type": "object",
          "properties": {
            "enabled": {
              "type": "boolean",
              "description": "Whether NER is enabled"
            },
            "serviceUrl": {
              "type": "string",
              "description": "NER service URL",
              "example": "http://localhost:5000/detect-pii"
            },
            "timeoutSeconds": {
              "type": "integer",
              "description": "Timeout for NER requests in seconds",
              "format": "int32",
              "example": 30
            },
            "maxSamples": {
              "type": "integer",
              "description": "Maximum samples to send to NER service",
              "format": "int32",
              "example": 100
            },
            "retryAttempts": {
              "type": "integer",
              "description": "Number of retry attempts for NER requests",
              "format": "int32",
              "example": 2
            },
            "circuitBreaker": {
              "$ref": "#/components/schemas/CircuitBreakerConfigDTO"
            }
          },
          "description": "Named Entity Recognition configuration settings"
        },
        "PoolConfigDTO": {
          "type": "object",
          "properties": {
            "connectionTimeout": {
              "type": "integer",
              "description": "Connection timeout in milliseconds",
              "format": "int64",
              "example": 30000
            },
            "idleTimeout": {
              "type": "integer",
              "description": "Idle timeout in milliseconds",
              "format": "int64",
              "example": 600000
            },
            "maxLifetime": {
              "type": "integer",
              "description": "Maximum connection lifetime in milliseconds",
              "format": "int64",
              "example": 1800000
            },
            "minimumIdle": {
              "type": "integer",
              "description": "Minimum idle connections",
              "format": "int32",
              "example": 5
            },
            "maximumPoolSize": {
              "type": "integer",
              "description": "Maximum pool size",
              "format": "int32",
              "example": 10
            }
          },
          "description": "Connection pool configuration settings"
        },
        "ReportingConfigDTO": {
          "type": "object",
          "properties": {
            "pdfEnabled": {
              "type": "boolean",
              "description": "Whether PDF report format is enabled"
            },
            "csvEnabled": {
              "type": "boolean",
              "description": "Whether CSV report format is enabled"
            },
            "textEnabled": {
              "type": "boolean",
              "description": "Whether text report format is enabled"
            },
            "reportOutputPath": {
              "type": "string",
              "description": "Path where reports are saved",
              "example": "./reports"
            }
          },
          "description": "Report generation configuration settings"
        },
        "DetectionRuleDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "Unique identifier of the rule",
              "example": "rule-123"
            },
            "name": {
              "type": "string",
              "description": "Name of the rule",
              "example": "Credit Card Detector"
            },
            "pattern": {
              "type": "string",
              "description": "Pattern used for detection",
              "example": "\\d{4}[- ]?\\d{4}[- ]?\\d{4}[- ]?\\d{4}"
            },
            "piiType": {
              "type": "string",
              "description": "PII type detected by this rule",
              "example": "CREDIT_CARD"
            },
            "confidenceScore": {
              "type": "number",
              "description": "Confidence score when this rule matches",
              "format": "double",
              "example": 0.95
            },
            "description": {
              "type": "string",
              "description": "Description of what the rule detects",
              "example": "Detects credit card numbers"
            },
            "enabled": {
              "type": "boolean",
              "description": "Whether this rule is active"
            },
            "ruleType": {
              "type": "string",
              "description": "Type of detection rule",
              "enum": [
                "REGEX",
                "HEURISTIC",
                "NER"
              ]
            }
          },
          "description": "PII detection rule configuration"
        },
...
        "Link": {
          "type": "object",
          "properties": {
            "href": {
              "type": "string"
            },
            "hreflang": {
              "type": "string"
            },
            "title": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "deprecation": {
              "type": "string"
            },
            "profile": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "templated": {
              "type": "boolean"
            }
          }
        },
...
        "SamplingRequest": {
          "required": [
            "columnName",
            "connectionId",
            "tableName"
          ],
          "type": "object",
          "properties": {
            "connectionId": {
              "type": "string",
              "format": "uuid"
            },
            "tableName": {
              "type": "string"
            },
            "columnName": {
              "type": "string"
            },
            "config": {
              "$ref": "#/components/schemas/SamplingConfigDTO"
            }
          }
        },
        "SamplingResponse": {
          "type": "object",
          "properties": {
            "meta": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            },
            "links": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Link"
              }
            },
            "tableName": {
              "type": "string"
            },
            "columnName": {
              "type": "string"
            },
            "sampleSize": {
              "type": "integer",
              "format": "int32"
            },
            "samplingMethod": {
              "type": "string"
            },
            "actualRowCount": {
              "type": "integer",
              "format": "int32"
            },
            "nullCount": {
              "type": "integer",
              "format": "int32"
            },
            "nullPercentage": {
              "type": "number",
              "format": "double"
            },
            "nonNullPercentage": {
              "type": "number",
              "format": "double"
            },
            "entropy": {
              "type": "number",
              "format": "double"
            },
            "entropyCalculated": {
              "type": "boolean"
            },
            "sampleValues": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "samplingTimeMs": {
              "type": "integer",
              "format": "int64"
            },
            "valueDistribution": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int64"
              }
            },
            "success": {
              "type": "boolean"
            }
          }
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "description": "Enter JWT token with Bearer prefix. Example: Bearer eyJhbGciOiJIUzI1NiIsIn...",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }